# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "学员端(hrwai_user) 学习/练习/考试/收藏/搜索/资料/论坛等 P0/P1 能力（ADR-0017/0018）。鉴权：`Authorization: Bearer \u003caccess JWT\u003e`（access 2h，`POST /api/auth/refresh` 换新双令牌）。响应统一 `{code,message,data}`（见 ADR-0005）。本页由 gin-swagger 生成，范围 B（~30 端点）。",
        "title": "叉车维修培训系统-学员端 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/chapter/{chapter_id}/slides": {
            "get": {
                "description": "公开访问，返回章节 PPT 转图片后的 slides",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学员端-课程"
                ],
                "summary": "章节幻灯片",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "章节ID",
                        "name": "chapter_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/forklift-training_pkg_response.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/forklift-training_internal_service.ChapterSlidesDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "章节不存在",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    }
                }
            }
        },
        "/chapter/{chapter_id}/slides/regenerate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "需登录，触发章节 PPT 重新转 slides（异步）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学员端-课程"
                ],
                "summary": "重新生成幻灯片",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "章节ID",
                        "name": "chapter_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/forklift-training_pkg_response.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/forklift-training_internal_service.ChapterSlidesDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    },
                    "404": {
                        "description": "章节不存在",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    }
                }
            }
        },
        "/course/{course_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "需登录，返回课程信息 + 章节 + 学员维度进度/是否已选/完成章节/最后位置（ADR-0017）；未登录时 last_* 为空",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学员端-课程"
                ],
                "summary": "课程详情（含学习进度）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "课程ID",
                        "name": "course_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/forklift-training_pkg_response.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/forklift-training_internal_service.CourseDetailDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    },
                    "404": {
                        "description": "课程不存在",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    }
                }
            }
        },
        "/course/{course_id}/chapter/{chapter_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "需登录，返回章节详情 + 相邻章节 + 学习状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学员端-课程"
                ],
                "summary": "章节详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "课程ID",
                        "name": "course_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "章节ID",
                        "name": "chapter_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/forklift-training_pkg_response.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/forklift-training_internal_service.ChapterDetailDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    },
                    "404": {
                        "description": "章节不存在",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    }
                }
            }
        },
        "/course/{course_id}/progress": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "需登录，上报章节学习时长/播放位置/完成态（ADR-0017）；body 支持 chapter_id/duration/duration_seconds/video_position/completed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学员端-课程"
                ],
                "summary": "上报学习进度",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "课程ID",
                        "name": "course_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "进度",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/forklift-training_pkg_response.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/forklift-training_internal_service.StudyProgressDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    }
                }
            }
        },
        "/courses": {
            "get": {
                "description": "公开访问，支持按专业方向 specialty_id / 等级 level_id 过滤，分页返回",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学员端-课程"
                ],
                "summary": "课程列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 12,
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "专业方向ID",
                        "name": "specialty_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "等级ID",
                        "name": "level_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/forklift-training_pkg_response.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/forklift-training_internal_service.CoursePageResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/student/courses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "学员已产生学习记录的课程列表（按最后学习时间倒序）+ continue_learning 置顶；包含封面/方向/等级/完成章节/最后位置（ADR-0017）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学员端"
                ],
                "summary": "我的课程",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/forklift-training_pkg_response.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/forklift-training_internal_service.StudentCoursesDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    }
                }
            }
        },
        "/student/courses/{course_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "指定课程的学习详情，包含每章进度/播放位置/完成状态（progress\u003e=100 为完成）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学员端"
                ],
                "summary": "单课程学习详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "课程ID",
                        "name": "course_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/forklift-training_pkg_response.R"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/forklift-training_internal_service.StudentCourseDetailDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    },
                    "404": {
                        "description": "课程不存在",
                        "schema": {
                            "$ref": "#/definitions/forklift-training_pkg_response.R"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "forklift-training_internal_service.ChapterDetailDTO": {
            "type": "object",
            "properties": {
                "chapter_id": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "content_type": {
                    "type": "string"
                },
                "course_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "file_url": {
                    "type": "string"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/forklift-training_internal_service.ChapterFileDTO"
                    }
                },
                "next_chapter_id": {
                    "type": "integer"
                },
                "order_num": {
                    "type": "integer"
                },
                "previous_chapter_id": {
                    "type": "integer"
                },
                "study_status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "forklift-training_internal_service.ChapterFileDTO": {
            "type": "object",
            "properties": {
                "chapter_id": {
                    "type": "integer"
                },
                "content_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "file_id": {
                    "type": "integer"
                },
                "file_name": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "file_url": {
                    "type": "string"
                }
            }
        },
        "forklift-training_internal_service.ChapterSlidesDTO": {
            "type": "object",
            "properties": {
                "chapter_id": {
                    "type": "integer"
                },
                "slides": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "forklift-training_internal_service.CourseDetailDTO": {
            "type": "object",
            "properties": {
                "chapters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/forklift-training_internal_service.ChapterDTO"
                    }
                },
                "completed_chapters": {
                    "type": "integer"
                },
                "course_info": {
                    "$ref": "#/definitions/forklift-training_internal_service.CourseDTO"
                },
                "is_enrolled": {
                    "type": "boolean"
                },
                "last_chapter_id": {
                    "type": "integer"
                },
                "last_position": {
                    "type": "integer"
                },
                "last_studied_at": {
                    "type": "string"
                },
                "progress": {
                    "type": "number"
                }
            }
        },
        "forklift-training_internal_service.CoursePageResult": {
            "type": "object",
            "properties": {
                "courses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/forklift-training_internal_service.CourseDTO"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pages": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "forklift-training_internal_service.StudentCourseDetailDTO": {
            "type": "object",
            "properties": {
                "chapters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/forklift-training_internal_service.StudentCourseChapterDTO"
                    }
                },
                "completed_chapters": {
                    "type": "integer"
                },
                "course_id": {
                    "type": "integer"
                },
                "course_name": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "last_chapter_id": {
                    "type": "integer"
                },
                "last_chapter_title": {
                    "type": "string"
                },
                "last_position": {
                    "type": "integer"
                },
                "last_studied_at": {
                    "type": "string"
                },
                "level_id": {
                    "type": "integer"
                },
                "progress": {
                    "type": "number"
                },
                "specialty_id": {
                    "type": "integer"
                },
                "study_duration": {
                    "type": "integer"
                },
                "total_chapters": {
                    "type": "integer"
                }
            }
        },
        "forklift-training_internal_service.StudentCoursesDTO": {
            "type": "object",
            "properties": {
                "continue_learning": {
                    "$ref": "#/definitions/forklift-training_internal_service.StudentCourseDTO"
                },
                "courses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/forklift-training_internal_service.StudentCourseDTO"
                    }
                }
            }
        },
...
        "forklift-training_internal_service.StudyProgressDTO": {
            "type": "object",
            "properties": {
                "completed_chapters": {
                    "type": "integer"
                },
                "progress": {
                    "type": "number"
                },
                "record_id": {
                    "type": "integer"
                },
                "study_duration": {
                    "type": "integer"
                }
            }
        },
...
        "forklift-training_pkg_response.R": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer \u003caccess JWT\u003e，示例：Bearer eyJhbGciOi...",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}